Move regression data from main file to dedicated file
 - Move FusionTable operations to the handler class
 - removed some asserts that were used when developing
