+++ mhcc/maintenance/mhcc_maintainer/mhcc_maintainer.py
@@ -5,6 +5,7 @@ initiates a resumable upload to handle the large datasetimport csv
 import csv
 import random
 import time
+from datetime import datetime
 
 from google_auth_oauthlib.flow import InstalledAppFlow
 from google.oauth2.credentials import Credentials
@@ -13,11 +14,14 @@ from services import HttpError
 from services import print_progress_bar as ppb
 from services import _write_as_csv as save
 
+STRTM_FMT = '%Y-%m-%dT%H:%M:%S.%f%z'
+
 LOCAL_KEYS = {}
 TABLE_LIST = {}
 
 SCOPES = ['https://www.googleapis.com/auth/fusiontables',
           'https://www.googleapis.com/auth/drive']
+
 
 def initialize(keys: dict, tables: dict):
     '''Read in the FusionTable IDs and any saved OAuth data/tokens
@@ -309,6 +313,7 @@ def prune_crowns(tableId: str):
     if not tableId or not isinstance(tableId, str):
         return
     # [Member, UID, LastSeen, LastCrown, LastTouched, B, S, G, MHCC, Squirrel]
+    return
 
 
 def keep_interesting_records(tableId: str):
@@ -443,6 +448,25 @@ def pick_table() -> str:
     return choice
 
 
+def addTimestringFields(records: list):
+    '''Add a naive-tz datetime to each annotated record for each millisecond-based time value (for easier human consumption)'''
+    transformables = set(('LastSeen', 'LastCrown', 'LastTouched', 'RankTime'))
+    keys = transformables.intersection(records[0].keys())
+    for row in records:
+        for k in keys:
+            ms = row[k]
+            dt = datetime.utcfromtimestamp(ms//1000).replace(microsecond=ms%1000*1000).strftime(STRTM_FMT)
+            row[f'{k}_TS'] = dt
+
+
+def clean_regressions(service: FusionTableHandler,
+                      time_start='2018-05-29T12:00:00.000000+0000',
+                      time_end='2018-11-30T12:00:00.000000+0000'):
+    from regression_fixer import clean_rank_regression, clean_crown_regression
+    uids = [x[1] for x in service.get_user_batch()]
+    args = (service, uids, time_start, time_end)
+    clean_rank_regression(*args, tableId=TABLE_LIST['MHCC Rank DB'])
+    clean_crown_regression(*args, tableId=TABLE_LIST['MHCC Crowns DB'])
 
 if __name__ == "__main__":
     initialize(LOCAL_KEYS, TABLE_LIST)
@@ -451,383 +475,8 @@ if __name__ == "__main__":
     handlers['FusionTables'].set_user_table(TABLE_LIST['MHCC Members'])
     #print('Pick a table')
     #table = pick_table()
-    table = '1xNi2C5Jfxz8QMkvVitUOgLumz3GTewm5t29hrkUF' # jacks ft id
     #print("Select the rank table")
     #prune_ranks(TABLE_LIST['MHCC Rank DB'], handlers['FusionTables'])
     #print('Select the crown table')
     #prune_crowns(TABLE_LIST['MHCC Crown DB'], handlers['FusionTables'])
-    # load jd from hdd
-    STRTM_FMT = '%Y-%m-%dT%H:%M:%S.%f%z'
-    ft = handlers['FusionTables']
-    users = ft.get_user_batch()
-    uids = [x[1] for x in users]
-
-    from datetime import datetime
-    def loadJacksData():
-        path = r'C:\Users\tehhowch\Downloads\MH Latest Crowns.csv'
-        with open(path, newline='') as file:
-            jacks_data = list(csv.reader(file))
-        jacks_cols = jacks_data.pop(0)
-        jd = [dict(zip(jacks_cols, [str(x[0]), int(x[1]) * 1000, int(x[2]), int(x[3]), int(x[4])] )) for x in jacks_data]
-        return jd
-
-    def getMHCCMembersInJacks(jd: dict):
-        mhcc_jd = [x for x in jd if x['snuid'] in uids]
-        for m in mhcc_jd:
-            name = [x[0] for x in users if x[1] == m['snuid']][0]
-            m['name'] = name
-        return mhcc_jd
-
-    def parseTimestamps(jd: dict):
-        '''Add a naive-tz datetime to each record (for easier human consumption)'''
-        for r in jd:
-            ms = r['timestamp']
-            dt = datetime.utcfromtimestamp(ms//1000).replace(microsecond=ms%1000*1000).strftime(STRTM_FMT)
-            r['LastSeen'] = ms
-            r['timestamp'] = dt
-
-    def parseMHCCTimes(mhcc: list):
-        '''Add a naive-tz datetime to each MHCC record for each MHCC millisecond time value (for easier human consumption)'''
-        transformables = set(['LastSeen', 'LastCrown', 'LastTouched', 'RankTime'])
-        keys = transformables.intersection(mhcc[0].keys())
-        for row in mhcc:
-            for k in keys:
-                ms = row[k]
-                dt = datetime.utcfromtimestamp(ms//1000).replace(microsecond=ms%1000*1000).strftime(STRTM_FMT)
-                row[f'{k}_TS'] = dt
-
-
-    def coerce_to_typed_info(tableId: str, data):
-        '''Converts str-only data elements to str, int, or float, in accordance with the FusionTable's
-        formatPattern and type for the given column'''
-        converter = get_column_mappings(tableId)
-        coerced = [{k: converter[k](v) for k, v in x.items()} for x in data]
-        return coerced
-
-    def get_rank_data(ranktime_start: str = None, ranktime_end: str = None, uid: str = None):
-        ''' For the given restrictions, obtain the corresponding MHCC Rank DB records '''
-        tid = TABLE_LIST['MHCC Rank DB']
-        sql = f'SELECT rowid, Member, UID, LastSeen, RankTime, Rank, \'MHCC Crowns\' FROM {tid}'
-        if any((ranktime_start, ranktime_end, uid)):
-            sql += ' WHERE '
-        # For any input time strings, convert to the corresponding UTC millis value.
-        if ranktime_start:
-            ts = datetime.strptime(ranktime_start, STRTM_FMT).timestamp() * 1000
-            sql += f'RankTime > {ts}'
-        if ranktime_end:
-            te = datetime.strptime(ranktime_end, STRTM_FMT).timestamp() * 1000
-            if ranktime_start:
-                sql += ' and '
-            sql += f'RankTime < {te}'
-        if uid:
-            if ranktime_start or ranktime_end:
-                sql += ' and '
-            sql += f'UID = {uid}'
-        sql += ' ORDER BY UID ASC, RankTime ASC'
-
-        # Download the subset of data from FusionTables
-        try:
-            rank_data = ft.get_query_result(query=sql, kb_row_size=0.25)
-        except HttpError:
-            print('Unable to obtain ROWIDs in bulk rank query')
-            rank_bytes = ft.query.sqlGet_media(sql=sql).execute()
-            rank_data = ft.bytestring_to_queryresult(rank_bytes)
-        # Convert from list of lists to list of dicts
-        headers = rank_data['columns']
-        output = (dict(zip(headers, x)) for x in rank_data['rows'])
-        return coerce_to_typed_info(tid, output)
-
-    def get_crown_data(crowntime_start: str = None, crowntime_end: str = None, uid: str = None):
-        ''' For the given restrictions, obtain the corresponding MHCC Crown DB records '''
-        tid = TABLE_LIST['MHCC Crowns DB']
-        sql = f'SELECT rowid, Member, UID, LastSeen, LastCrown, LastTouched, Bronze, Silver, Gold, MHCC, Squirrel FROM {tid}'
-        if any((crowntime_start, crowntime_end, uid)):
-            sql += ' WHERE '
-        # For any input time strings, convert to the corresponding UTC millis value.
-        if crowntime_start:
-            ts = datetime.strptime(crowntime_start, STRTM_FMT).timestamp() * 1000
-            sql += f'LastTouched > {ts}'
-        if crowntime_end:
-            te = datetime.strptime(crowntime_end, STRTM_FMT).timestamp() * 1000
-            if crowntime_start:
-                sql += ' and '
-            sql += f'LastTouched < {te}'
-        if uid:
-            if crowntime_start or crowntime_end:
-                sql += ' and '
-            sql += f'UID = {uid}'
-        sql += ' ORDER BY UID ASC, LastTouched ASC'
-
-        # Download the subset of data from FusionTables
-        try:
-            crown_data = ft.get_query_result(query=sql, kb_row_size=0.25)
-        except HttpError:
-            print('Unable to obtain ROWIDs in bulk crown query')
-            crown_bytes = ft.query.sqlGet_media(sql=sql).execute()
-            crown_data = ft.bytestring_to_queryresult(crown_bytes)
-        # Convert from list of lists to list of dicts
-        headers = crown_data['columns']
-        output = (dict(zip(headers, x)) for x in crown_data['rows'])
-        return coerce_to_typed_info(tid, output)
-
-
-    def get_column_mappings(id):
-        '''Query the table and determine the appropriate str/int/float type coercion for each column.'''
-        def get_as_int(val):
-            try:
-                return int(val)
-            except ValueError:
-                try:
-                    return int(float(val))
-                except ValueError as err:
-                    if val == 'NaN':
-                        return None
-                    raise err
-
-        cols = ft.table.get(tableId=id, fields='columns(columnId,name,type,formatPattern)').execute()['columns']
-        maps = {'rowid': str}
-        for c in cols:
-            patt = c['formatPattern']
-            cType = c['type'] # NUMBER or STRING (in future, maybe DATETIME)
-            if cType == 'NUMBER':
-                if patt == 'NUMBER_INTEGER':
-                    f_func = get_as_int
-                else:
-                    f_func = float
-            else:
-                f_func = str
-            assert c['name'] not in maps # column name must be unique
-            maps[c['name']] = f_func
-        return maps
-
-    def deannotate(headers, record):
-        ''' Convert the input record back to a list of lists '''
-        output = []
-        for col in headers:
-            output.append(record[col])
-        return output
-
-
-    def sort_by_ranktime(record):
-        return record['RankTime']
-    def sort_by_lasttouched(record):
-        return record['LastTouched']
-
-    def get_first_decrease(records: list, key: str):
-        ''' Find the first record for the given UID in which the value of the given key decreases '''
-        if records:
-            last_key_value = None
-            start = 0
-            while last_key_value is None and start < len(records):
-                last_key_value = records[start][key]
-                start += 1
-            if last_key_value is None:
-                return None
-            for i, record in enumerate(records[start:], start):
-                current_key_value = record[key]
-                if(last_key_value > current_key_value):
-                    return i
-                else:
-                    last_key_value = current_key_value
-        return None
-
-    def get_first_correct(ranks: list, key: str, start: int):
-        ''' Find the first record for the given UID in which the value of the given key returns to "normal" '''
-        if ranks:
-            assert start < len(ranks)
-            assert start > 0
-            last_key_value = ranks[start - 1][key]
-            for i, record in enumerate(ranks[start:], start):
-                if record[key] >= last_key_value:
-                    return i
-        return None
-
-    def get_prune_range(sorted_records: list):
-        ''' Return a tuple with the range of the bad data in the input list '''
-        # Find the first record where LastSeen decreases
-        first_bad = get_first_decrease(sorted_records, 'LastSeen')
-        if first_bad is None:
-            return None
-        # Find the next record that restores the required "LastSeen increasing" property
-        first_good = get_first_correct(sorted_records, 'LastSeen', first_bad)
-        if first_good is None:
-            print(f'All records after {first_bad} are bad ({len(sorted_records)} total records).')
-        else:
-            assert first_good < len(sorted_records), f'First good index exceeds allowable dimension'
-        return (first_bad, first_good)
-
-
-    def delete_records_by_rowid(service: FusionTableHandler, tableId: str, rowids: list):
-        ''' Delete the given records from the given FusionTable. Does not back up the table first.'''
-        raw_sql = 'DELETE FROM ' + tableId + ' WHERE ROWID IN ({})'
-        max_query_length = service.MAX_GET_QUERY_LENGTH * .75 - 2 * max(len(str(x)) for x in rowids)
-        deleted = 0
-        while rowids:
-            query_ids = [rowids.pop()]
-            while rowids and len(raw_sql.format(','.join(query_ids))) < max_query_length:
-                query_ids.append(rowids.pop())
-            query = raw_sql.format(','.join(query_ids))
-            assert len(query) <= service.MAX_GET_QUERY_LENGTH, f'Query too long {len(query)} vs {service.MAX_GET_QUERY_LENGTH}'
-            print(f'Estimated remaining time of {len(rowids) / len(query_ids)} sec.')
-            resp = service.query.sql(sql=query).execute(num_retries=2)
-            deleted += int(resp['rows'][0][0])
-            time.sleep(1)
-        print(f'Deleted {deleted} rows.')
-        return deleted
-
-    def delete_records_by_criteria(service: FusionTableHandler, tableId: str, records):
-        ''' Delete the given records by specifying various criteria other than rowid.
-        E.g. all bad records for a given UID have the same LastSeen and fall in a certain range of RankTimes / LastTouched'''
-        pass
-
-
-    at_risk = getMHCCMembersInJacks(loadJacksData())
-    time_start = '2018-05-29T12:00:00.000000+0000'
-    time_end = '2018-11-30T12:00:00.000000+0000'
-    ranks = get_rank_data(time_start, time_end)
-    #parseMHCCTimes(ranks)
-    problem_uids = set(x['snuid'] for x in at_risk)
-    print(f'Indexing {len(ranks)} by UID')
-    indexed_ranks = {}
-    for record in ranks:
-        indexed_ranks.setdefault(record['UID'], []).append(record)
-
-    collected_bad_ranks = []
-    for i, uid in enumerate(uids):
-        member_ranks = indexed_ranks.get(uid, [])
-
-        member_ranks.sort(key=sort_by_ranktime)
-        indices = get_prune_range(member_ranks)
-        while indices is not None:
-            # Verify assumption that only members with entries in Jack's Crown DB have bad data
-            assert uid in problem_uids, f'{uid} is not in Jacks db'
-            to_prune = member_ranks[indices[0] : indices[1]]
-            assert len(set(x['LastSeen'] for x in to_prune)) == 1
-            collected_bad_ranks.extend(to_prune)
-            # It is possible there is more than one set of bad data. Remove all of them.
-            del member_ranks[indices[0] : indices[1]]
-            indices = get_prune_range(member_ranks)
-        # Update the stored data to reflect the fixed representation.
-        indexed_ranks[uid] = member_ranks
-
-    indexed_bad_data = {}
-    if collected_bad_ranks:
-        # Write this data to disk (allow avoiding an expensive requery of the table)
-        with open('bad_rank_data.csv', 'w', encoding='utf-8', newline='') as file:
-            dw = csv.DictWriter(file, fieldnames=list(collected_bad_ranks[0].keys()), quoting=csv.QUOTE_NONNUMERIC)
-            dw.writeheader()
-            dw.writerows(collected_bad_ranks)
-
-        min_ms = min(x["RankTime"] for x in collected_bad_ranks if x['MHCC Crowns'] > 0)
-        print(f'Earliest rank regression was on {datetime.utcfromtimestamp(min_ms//1000).replace(microsecond=min_ms%1000*1000).strftime(STRTM_FMT)}')
-
-        # Create a backup of the rank table
-        tid = TABLE_LIST['MHCC Rank DB']
-        backup = ft.backup_table(tid)
-        if backup:
-            # Avoid deletions while the backup is cloning
-            while True:
-                tasks = ft.get_tasks(backup['tableId'])
-                if tasks:
-                    print(tasks[0]['type'], tasks[0]['progress'])
-                    time.sleep(5)
-                else:
-                    break
-
-            rowids = [x['rowid'] for x in collected_bad_ranks]
-            # Delete the rows with the given ROWIDs
-            num_rows = ft.count_rows(tid)
-            target_rows = num_rows - len(rowids)
-            delete_records_by_rowid(ft, tid, rowids)
-            new_count = ft.count_rows(tid)
-            if new_count >= num_rows:
-                print('Rank table does not have fewer rows', new_count, num_rows)
-            if new_count != target_rows:
-                print(f'Expected rank table to have {target_rows}, but it has {new_count}')
-        else:
-            print('Skipped rank data deletion due to failed backup')
-
-    # Bad data may have additionally accumulated in the Crowns DB that does not quite correspond to that visible via the Rank DB
-    # For example, if information from all data sources is added, then the recorded information toggles between the two, but only the most
-    # recently added would be presented for inclusion in the Rank DB.
-    print(f'Collecting crown data in range {time_start}-{time_end}')
-    crowns = get_crown_data(time_start, time_end)
-    crown_header_order = [x['name'] for x in ft.get_all_columns(TABLE_LIST['MHCC Crowns DB'])['columns']]
-    collected_bad_crowns = []
-    lastcrown_recalculations = []
-    indexed_crowns = {}
-    for record in crowns:
-        indexed_crowns.setdefault(record['UID'], []).append(record)
-    for i, uid in enumerate(uids):
-        member_crowns = indexed_crowns.get(uid, [])
-        member_crowns.sort(key=sort_by_lasttouched)
-        lastcrown_modifications = []
-
-        indices = get_prune_range(member_crowns)
-        while indices is not None:
-            # Verify assumption that only members with entries in Jack's Crown DB have bad data
-            assert uid in problem_uids, f'{uid} is not in Jacks db'
-            to_prune = member_crowns[indices[0] : indices[1]]
-            assert len(set(x['LastSeen'] for x in to_prune)) == 1
-            collected_bad_crowns.extend(to_prune)
-            # Recalculate the "LastCrown" column
-            if indices[1] is not None:
-                # There is a record to compute with.
-                reference = member_crowns[indices[0] - 1]
-                modified = member_crowns[indices[1]]
-                if all(modified[key] == reference[key] for key in ('Silver', 'Gold', 'MHCC')):
-                    modified['LastCrown'] = reference['LastCrown']
-                    lastcrown_modifications.append({'rowid': modified['rowid'],
-                                                    'new_record': deannotate(crown_header_order, modified)})
-                elif modified['LastCrown'] != modified['LastSeen']:
-                    print(f'Has new MHCC crowns but not new LastCrown. Updating from {modified["LastCrown"]} to {modified["LastSeen"]}')
-                    modified['LastCrown'] = modified['LastSeen']
-                    lastcrown_modifications.append({'rowid': modified['rowid'],
-                                                    'new_record': deannotate(crown_header_order, modified)})
-            # It is possible there is more than one set of bad data. Remove all of them.
-            del member_crowns[indices[0] : indices[1]]
-            indices = get_prune_range(member_crowns)
-        # Update the stored data to reflect the fixed representation.
-        indexed_crowns[uid] = member_crowns
-        kept_rowids = set(x['rowid'] for x in member_crowns)
-        lastcrown_recalculations.extend([x for x in lastcrown_modifications if x['rowid'] in kept_rowids])
-    if collected_bad_crowns:
-        # Write this data to disk (allow avoiding an expensive requery of the table)
-        with open('bad_crown_data.csv', 'w', encoding='utf-8', newline='') as file:
-            dw = csv.DictWriter(file, fieldnames=list(collected_bad_crowns[0].keys()), quoting=csv.QUOTE_NONNUMERIC)
-            dw.writeheader()
-            dw.writerows(collected_bad_crowns)
-
-        min_ms = min(x["LastTouched"] for x in collected_bad_crowns if x['MHCC'] > 0)
-        print(f'Earliest data regression was on {datetime.utcfromtimestamp(min_ms//1000).replace(microsecond=min_ms%1000*1000).strftime(STRTM_FMT)}')
-        
-        # Create a backup of the crowns table
-        tid = TABLE_LIST['MHCC Crowns DB']
-        backup = ft.backup_table(tid)
-        if backup:
-            # Avoid deletions while the backup is cloning
-            while True:
-                tasks = ft.get_tasks(backup['tableId'])
-                if tasks:
-                    print(tasks[0]['type'], tasks[0]['progress'])
-                    time.sleep(5)
-                else:
-                    break
-
-            rowids = [x['rowid'] for x in collected_bad_crowns]
-            # Delete the rows with the given ROWIDs
-            num_rows = ft.count_rows(tid)
-            target_rows = num_rows - len(rowids)
-            delete_records_by_rowid(ft, tid, rowids)
-            new_count = ft.count_rows(tid)
-            if new_count >= num_rows:
-                print('Crown table does not have fewer rows', new_count, num_rows)
-            if new_count != target_rows:
-                print(f'Expected crown table to have {target_rows}, but it has {new_count}')
-
-            # Update the associated LastCrown records
-            rowids = [x['rowid'] for x in lastcrown_recalculations]
-            delete_records_by_rowid(ft, tid, rowids)
-            added = ft.import_rows(tid, [x['new_record'] for x in lastcrown_recalculations])
-            print(f'{added} of {len(lastcrown_recalculations)} rows with corrected LastCrown values were uploaded.')
     print('waiting for you to do stuff')
